JuDO

Documentation for JuDO.

JuDO.DynModel

JuDO.integral

JuDO.objective_has_measures

JuDO.optimize

JuDO.set_interpolant

JuDO.set_intervals

JuDO.to_NDF

JuDO.value

JuMP.build_variable

JuDO.DynModel — Method

DynModel <: Abstract_Dynamic_Model

Storing, displaying the information of the model

JuDO.integral — Method

JuMP.integral(expr::JuMP.AbstractJuMPScalar, phase::Int) -> Integral

Creates an integral representation for the given expression and phase.

JuDO.objective_has_measures — Method

objective_has_measures(model::JuMP.Model) -> Bool

Returns true if the objective function contains any measure variables.

JuDO.optimize — Method

optimize!(m::JuMP.Model)
Optimize the model.

JuDO.set_interpolant — Method

set_interpolant(m::DynModel, interpolant::DOI.AbstractDynamicSolution, var::DynamicVarRef)
set the interpolant for a dynamic variable in the model

JuDO.set_intervals — Method

set_intervals(m::JuMP.Model, intervals::AbstractIntervals)
Set the intervals for the model.

JuDO.to_NDF — Function

to_NDF(f::AbstractDynamicFunction)

Convert an AbstractDynamicFunction into its NonlinearDynamicFunction equivalent.

This function builds a nested expression tree to represent the input function.

JuDO.value — Method

value(m::JuMP.Model, var::DynamicVarRef)
Get the value of a dynamic variable.

JuMP.build_variable — Method

Extending Jump.build_variable